Add ability to replace objects when using copy and paste #2620
